.github/requirements.txt
google-generativeai
requests

